结构稍作变动

+++ ORM/Interface/IInsert/IInsert.cs
@@ -4,7 +4,7 @@ namespace ORM.Interface
 {
     public interface IInsert<in T>
     {
-        int Insert(T model, Transaction transaction = null);
-        int InsertBatch(IEnumerator<T> models, Transaction transaction = null);
+        long Insert(T model, Transaction transaction = null);
+        long InsertBatch(IEnumerator<T> models, Transaction transaction = null);
     }
 }
